requirements.txt
boto==2.38.0
PyYAML==3.11
wheel==0.24.0
MySQL-python==1.2.3